Tidy up.
- Remove unused code
- Regroup glyph info var allocation accessors
- Move (Would)ApplyContext into separate file
