+++ harfbuzz/src/hb-aat-layout-morx-table.hh
@@ -29,7 +29,6 @@
 
 #include "hb-open-type.hh"
 #include "hb-aat-layout-common.hh"
-#include "hb-ot-layout-common.hh"
 #include "hb-aat-map.hh"
 
 /*
