+++ harfbuzz/src/hb-ot-layout-common.hh
@@ -41,35 +41,6 @@
 #ifndef RB_MAX_CONTEXT_LENGTH
 #define RB_MAX_CONTEXT_LENGTH 64
 #endif
-#ifndef RB_CLOSURE_MAX_STAGES
-/*
- * The maximum number of times a lookup can be applied during shaping.
- * Used to limit the number of iterations of the closure algorithm.
- * This must be larger than the number of times add_pause() is
- * called in a collect_features call of any shaper.
- */
-#define RB_CLOSURE_MAX_STAGES 32
-#endif
-
-#ifndef RB_MAX_SCRIPTS
-#define RB_MAX_SCRIPTS 500
-#endif
-
-#ifndef RB_MAX_LANGSYS
-#define RB_MAX_LANGSYS 2000
-#endif
-
-#ifndef RB_MAX_FEATURES
-#define RB_MAX_FEATURES 750
-#endif
-
-#ifndef RB_MAX_FEATURE_INDICES
-#define RB_MAX_FEATURE_INDICES 1500
-#endif
-
-#ifndef RB_MAX_LOOKUP_INDICES
-#define RB_MAX_LOOKUP_INDICES 20000
-#endif
 
 namespace OT {
 
@@ -714,194 +685,6 @@ public:
 };
 
 /*
- * Item Variation Store
- */
-
-struct VarRegionAxis
-{
-    float evaluate(int coord) const
-    {
-        int start = startCoord, peak = peakCoord, end = endCoord;
-
-        /* TODO Move these to sanitize(). */
-        if (unlikely(start > peak || peak > end))
-            return 1.;
-        if (unlikely(start < 0 && end > 0 && peak != 0))
-            return 1.;
-
-        if (peak == 0 || coord == peak)
-            return 1.;
-
-        if (coord <= start || end <= coord)
-            return 0.;
-
-        /* Interpolate */
-        if (coord < peak)
-            return float(coord - start) / (peak - start);
-        else
-            return float(end - coord) / (end - peak);
-    }
-
-    bool sanitize(rb_sanitize_context_t *c) const
-    {
-        return c->check_struct(this);
-        /* TODO Handle invalid start/peak/end configs, so we don't
-         * have to do that at runtime. */
-    }
-
-public:
-    F2DOT14 startCoord;
-    F2DOT14 peakCoord;
-    F2DOT14 endCoord;
-
-public:
-    DEFINE_SIZE_STATIC(6);
-};
-
-struct VarRegionList
-{
-    float evaluate(unsigned int region_index, const int *coords, unsigned int coord_len) const
-    {
-        if (unlikely(region_index >= regionCount))
-            return 0.;
-
-        const VarRegionAxis *axes = axesZ.arrayZ + (region_index * axisCount);
-
-        float v = 1.;
-        unsigned int count = axisCount;
-        for (unsigned int i = 0; i < count; i++) {
-            int coord = i < coord_len ? coords[i] : 0;
-            float factor = axes[i].evaluate(coord);
-            if (factor == 0.f)
-                return 0.;
-            v *= factor;
-        }
-        return v;
-    }
-
-    bool sanitize(rb_sanitize_context_t *c) const
-    {
-        return c->check_struct(this) && axesZ.sanitize(c, (unsigned int)axisCount * (unsigned int)regionCount);
-    }
-
-protected:
-    HBUINT16 axisCount;
-    HBUINT16 regionCount;
-    UnsizedArrayOf<VarRegionAxis> axesZ;
-
-public:
-    DEFINE_SIZE_ARRAY(4, axesZ);
-};
-
-struct VarData
-{
-    unsigned int get_region_index_count() const
-    {
-        return regionIndices.len;
-    }
-
-    unsigned int get_row_size() const
-    {
-        return shortCount + regionIndices.len;
-    }
-
-    unsigned int get_size() const
-    {
-        return itemCount * get_row_size();
-    }
-
-    float get_delta(unsigned int inner, const int *coords, unsigned int coord_count, const VarRegionList &regions) const
-    {
-        if (unlikely(inner >= itemCount))
-            return 0.;
-
-        unsigned int count = regionIndices.len;
-        unsigned int scount = shortCount;
-
-        const HBUINT8 *bytes = get_delta_bytes();
-        const HBUINT8 *row = bytes + inner * (scount + count);
-
-        float delta = 0.;
-        unsigned int i = 0;
-
-        const HBINT16 *scursor = reinterpret_cast<const HBINT16 *>(row);
-        for (; i < scount; i++) {
-            float scalar = regions.evaluate(regionIndices.arrayZ[i], coords, coord_count);
-            delta += scalar * *scursor++;
-        }
-        const HBINT8 *bcursor = reinterpret_cast<const HBINT8 *>(scursor);
-        for (; i < count; i++) {
-            float scalar = regions.evaluate(regionIndices.arrayZ[i], coords, coord_count);
-            delta += scalar * *bcursor++;
-        }
-
-        return delta;
-    }
-
-    void get_scalars(const int *coords,
-                     unsigned int coord_count,
-                     const VarRegionList &regions,
-                     float *scalars /*OUT */,
-                     unsigned int num_scalars) const
-    {
-        unsigned count = rb_min(num_scalars, regionIndices.len);
-        for (unsigned int i = 0; i < count; i++)
-            scalars[i] = regions.evaluate(regionIndices.arrayZ[i], coords, coord_count);
-        for (unsigned int i = count; i < num_scalars; i++)
-            scalars[i] = 0.f;
-    }
-
-    bool sanitize(rb_sanitize_context_t *c) const
-    {
-        return c->check_struct(this) && regionIndices.sanitize(c) && shortCount <= regionIndices.len &&
-               c->check_range(get_delta_bytes(), itemCount, get_row_size());
-    }
-
-protected:
-    const HBUINT8 *get_delta_bytes() const
-    {
-        return &StructAfter<HBUINT8>(regionIndices);
-    }
-
-    HBUINT8 *get_delta_bytes()
-    {
-        return &StructAfter<HBUINT8>(regionIndices);
-    }
-
-protected:
-    HBUINT16 itemCount;
-    HBUINT16 shortCount;
-    ArrayOf<HBUINT16> regionIndices;
-    /*UnsizedArrayOf<HBUINT8>bytesX;*/
-public:
-    DEFINE_SIZE_ARRAY(6, regionIndices);
-};
-
-struct VariationStore
-{
-    float get_delta(unsigned int outer, unsigned int inner, const int *coords, unsigned int coord_count) const
-    {
-        if (unlikely(outer >= dataSets.len))
-            return 0.f;
-
-        return (this + dataSets[outer]).get_delta(inner, coords, coord_count, this + regions);
-    }
-
-    bool sanitize(rb_sanitize_context_t *c) const
-    {
-        return c->check_struct(this) && format == 1 && regions.sanitize(c, this) && dataSets.sanitize(c, this);
-    }
-
-protected:
-    HBUINT16 format;
-    LOffsetTo<VarRegionList> regions;
-    LOffsetArrayOf<VarData> dataSets;
-
-public:
-    DEFINE_SIZE_ARRAY(8, dataSets);
-};
-
-/*
  * Feature Variations
  */
 
@@ -1095,32 +878,6 @@ protected:
 
 public:
     DEFINE_SIZE_ARRAY_SIZED(8, varRecords);
-};
-
-/*
- * Device Table
- */
-
-extern "C" {
-RB_EXTERN rb_position_t rb_device_get_x_delta(const char *data, const rb_font_t *font);
-RB_EXTERN rb_position_t rb_device_get_y_delta(const char *data, const rb_font_t *font);
-}
-
-struct Device
-{
-    rb_position_t get_x_delta(rb_font_t *font) const
-    {
-        return rb_device_get_x_delta((const char*)this, font);
-    }
-    rb_position_t get_y_delta(rb_font_t *font) const
-    {
-        return rb_device_get_y_delta((const char*)this, font);
-    }
-
-    bool sanitize(rb_sanitize_context_t *c) const
-    {
-        return true;
-    }
 };
 
 } /* namespace OT */
