+++ src/ffi.rs
@@ -289,24 +289,11 @@ extern "C" {
 
     pub fn rb_ot_apply_context_get_nesting_level_left(ctx: *const rb_ot_apply_context_t) -> u32;
 
-    pub fn rb_ot_apply_context_get_has_glyph_classes(ctx: *const rb_ot_apply_context_t) -> rb_bool_t;
-
     pub fn rb_ot_apply_context_get_auto_zwnj(ctx: *const rb_ot_apply_context_t) -> rb_bool_t;
 
     pub fn rb_ot_apply_context_get_auto_zwj(ctx: *const rb_ot_apply_context_t) -> rb_bool_t;
 
     pub fn rb_ot_apply_context_get_random(ctx: *const rb_ot_apply_context_t) -> rb_bool_t;
-
-    pub fn rb_ot_apply_context_gdef_mark_set_covers(
-        ctx: *const rb_ot_apply_context_t,
-        set_index: u32,
-        glyph_id: rb_codepoint_t,
-    ) -> rb_bool_t;
-
-    pub fn rb_ot_apply_context_gdef_get_glyph_props(
-        ctx: *const rb_ot_apply_context_t,
-        glyph_id: rb_codepoint_t,
-    ) -> u32;
 
     pub fn rb_ot_apply_context_random_number(ctx: *mut rb_ot_apply_context_t) -> u32;
